MQTT: Fixed MQTT header constructing and parsing
Bit fields must not be used in communication protocol frame parsing.
See http://stackoverflow.com/questions/1490092/c-c-force-bit-field-order-and-alignment.

The bit-fields used in AWS C library caused issues in PowerPC target I'm using.

This might not be the most beautiful way to fix it, but it works. Tested on x86
and PowerPC architectures.

fixes #59