support 'certificates in sources' option
Some embedded devices must have the certificate files in their source
when they are built. If you disable the 'MBEDTLS_FS_IO' option of
mbedTLS, you can not use the function to import files. This patch
includes a script that inserts the certificate into the source code.
Here's how to use the script:

$ cd ./certs
$ sh mkcertData.sh <root CA file> <certificate file> <private key file>

The certData.c file is generated, and this patch contains mbedtls
function substitution to use it.

Signed-off-by: Junhwan Park <junhwan.park@samsung.com>