Updated cbmc proof instructions in cbmc-batch.yaml.
A file cbmc-batch.yaml in each proof directory under cbmc/proofs gives
instructions for how to recheck the proof in that directory with CBMC.
The proofs are rechecked each time a commit or pull request is added
to the repository.  A major commit changed the names of the loop
labels CBMC uses (by changing the names of the functions containing
the loops).  So the CBMC proof instructions had to be updated, too, by
running 'make cbmc-batch.yaml' in that directory.