Clone TinyCBOR for CBMC proofs in continuous integration (#382)
The build flow for the CBMC proofs does a "git submodule init" and
"git submodule update" to pull in the TinyCBOR submodule before
building MQTT for CBMC.  In continuous integration, however, we pull
down a straight tar file of the commit and don't actually clone the
repository, and the CBMC build flow fails at "git submodule init" in
continuous integration with the error "not a git repository".  This
pull request just manually clones the TinyCBOR into place if it is
missing.