Don't treat resubscribe failures as reconnect failures (#548)
The auto-reconnect sequence comprises of connect followed
by subscribe to the topics the client was subscribed to before
disconnect. A failed subscribe after a successful connect during
auto reconnect sequence used to re-trigger the whole auto
reconnect sequence (connect followed by subscribes). 

There is no need to attempt to connect on a subscribe failure
as the client is already connected. This change modifies this
behavior so that a failed subscribe does not attempt to connect
again and in the next iteration of yield only the remaining
subscribes are attempted.