Automate generation of cbmc-batch.yaml during continuous integration. (#690)
The CBMC proof checking done during continuous integration is driven
by a set of yaml files.  Proofs are developed locally using a Makefile,
but proofs are checked in CI using a file cbmc-batch.yaml sitting next
to the Makefile in the proof directory.

The Makefile has a target to generate the yaml file from the Makefile.
This patch automates the regeneration of the yaml files during CI
before launching the proof checking.