Fix retrieving last string in indefinite length containers
Indefinite length containers have a break character (0xFF) at the end
to indicate the end of the container. When the last element of an
indefinite length container was a string, the returned string was offset
by one because of the break character.

This commit addresses the problem by adjusting the string to account for
the break character in case the string happens to be the last element in
an indefinite length container.

Signed-off-by: Gaurav Aggarwal <aggarg@amazon.com>