Ci/coverity integration (#967)
* Add misra config

Add the misra configration for coverity. This is taken from the V4_Beta
branch and will act as a living reference to what is checked.

* Explicitly link DL and PThreads libraries

The build node in CI is unable to find PTHREADS and DL libraries.

* Move thread related options to proper file

Move the thread related options from the root CMakeLists.txt file to the
most specific one (the TLS demo)

* Add explanatory comment to tls demo.

Add an explanatory comment to the TLS demo. Ordinarily the demo does not
use threading or dynamic loading, but OpenSSL requires them in CI and
the CI platform requires that they are explicitly linked.