Add reconnection logic to remaining HTTP demos (#1280)
This adds the same reconnection logic added in PR #1253 to the remaining HTTP demos. Along with this, the function connectToServerWithBackoffRetries is refactored into a common place under demos/http/common. The ports are also renamed differently for all the demos so that they can be passed as CFLAGS, and the dependency on httpbin.org is removed.