Lower CMake minimum required version from 3.13.0 to 3.2.0 (#1306)
With these small changes, we can support CMake 3.2.0 or higher. Unfortunately, 3.1.0 has a bug in finding OpenSSL libraries, but CMake 3.2.0 has been tested to work for both demos and tests. This change is especially useful for people with older Linux distros with package managers that don't install the latest CMake.