Build tests without also building demos and add missing log (#1428)
Demos were also added as targets when the BUILD_TESTS flag was activated. However, when tests are built, there is usually no intention to build demos.

In addition, the preprocessor logic for checking if certain macros are defined is updated to also show a message when all macros are found instead of only when a macro is missing.