Support hostname verification and partial writes in Openssl wrapper (#1433)
When the SNI hostname is set, we need to use SSL_set1_host to verify the hostname against the server's certificate. Also, when the payload was too large, OpenSSL would just return SSL_WANT_WRITE instead of trying to write as many bytes as it could, potentially causing an infinite loop in libraries. The SSL_MODE_ENABLE_PARTIAL_WRITE flag allows OpenSSL to support partial writes so that less bytes than the expected payload can be returned.