Add GitHub Actions job to upload release (#1503)
Add a GitHub Actions workflow to automate creation of release and uploading of release asset for an existing tag after the execution of the Create Tag And Zip workflow.
The job takes the following input parameters:

The name of the workflow that creates the tag and the ZIP artifact. Currently, as the tag-and-zip.yml workflow does the work, this is the default value of the field.
The run ID of the above workflow to source the ZIP artifact from and upload as release asset for the tag.
Existing tag in repository whose release will be created.