Add CI check for manifest.yml (#1677)
This PR adds a mechanism to ensure that the manifest.yml file always represent the latest information of the repository. The check verifies that the file contains entries for all required submodules of the repository as well as those submodule entries represent the latest commit ID present of the main branch. This prevents the manifest.yml file from containing release tags when library submodule pointers are updated to represent non-release tags.

The CI check is executed by calling the manifest-verifier GitHub Action in FreeRTOS/CI-CD-GitHub-Actions repository.