posix: Skip setting socket send and receive timeout if unimplemented (#1743)
* posix: Skip setting socket send and receive timeout if unimplemented

According to POSIX spec for setsockopt [1], SO_SNDTIMEO and SO_RCVTIMEO
options at the socket level are optional. A complaint implementation
shall set the errno to ENOPROTOOPT if they are not supported.

Currently Sockets_Connect() treats such case as an error and aborts the
connection. Update the codes to check the errno against ENOPROTOOPT and
skip it when unsupported.

[1] https://pubs.opengroup.org/onlinepubs/000095399/functions/setsockopt.html

Signed-off-by: Yiting Wang <yiting.wang@windriver.com>