Avoid hardcoded linking to libpthread
Currently ota_demo_core_http and ota_demo_core_mqtt make rules hardcode
linking to libpthread but VxWorks does not provide a separate pthread
library as a regular file, instead it's the libc that provides all the
pthread APIs.

We can certainly add the following logic in the CMakeLists.txt:

  if(VXWORKS)
      # remove pthread from the linking libraries
  else()
      # original linking libraries with pthread
  endif()

but that does not look very elegant. In fact, we don't have to use
VXWORKS to create a separate branch, the find_package(Threads) has
already provided Threads:Threads IMPORTED target [1] that holds the
thread library if found.

However by default the IMPORTED target does not have global scope
which means the target is only visible in the platform directory,
where find_package(Threads) is called. It's not possible to use it
in the demo directory where we want to replace the hardcoded pthread.
CMake 3.11 introduced the ability to promote an imported target to
global visibility by setting the target's IMPORTED_GLOBAL [2] property
to true, but that requires us to bump up the CMake minimum version
from current 3.2.0 to 3.11.0, which is not ideal as Ubuntu 18.04, a
still popular development host OS nowadays, is shipping CMake 3.10.

To keep current CMake minimum version requirement while promoting
the IMPORTED target scope to global, we need to move the call to
find_package(Threads) in the root directory. With that change, the
creation of Threads_FOUND as a cache variable is no longer needed.

[1] https://cmake.org/cmake/help/latest/module/FindThreads.html
[2] https://cmake.org/cmake/help/latest/prop_tgt/IMPORTED_GLOBAL.html

Signed-off-by: Yiting Wang <yiting.wang@windriver.com>