Remove useless check for valid certificate
If certificate check fails, SSL_Connect will fail, thus the removed code will
not run. It would only runs when successful. That function is meant to check the
error code when there is a certificate validation issue.

SSL_get_verify_result is being used incorrectly here; it is intended to get
the validation error reason when SSL_Connect fails, and that failure is due to
an invalid certificate.